refactor: 그룹 참여 페이지 내부 컴포넌트 추상화
- GroupInitPage 컴포넌트 추상화를 통해, 페이지 내부 레이아웃이 잘 파악되도록 수정
