+++ frontend/src/pages/GroupInitPage/GroupInitPage.styles.tsx
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+
+const StyledContainer = styled.div`
+  height: 100vh;
+  width: 100vw;
+`;
+
+export { StyledContainer };
